running-coach-bot/requirements.txt
anthropic>=0.49.0
python-telegram-bot>=21.0
requests>=2.31.0
python-dotenv>=1.0.0
